feat: extract fish image map to utils to fix Fast Refresh warning
